+++ source/MRMesh/MRMesh.vcxproj.filters
@@ -1294,6 +1294,9 @@
     <ClInclude Include="MRParallelMinMax.h">
       <Filter>Source Files\Basic</Filter>
     </ClInclude>
+    <ClInclude Include="MRMinMaxArg.h">
+      <Filter>Source Files\Basic</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="MRParallelProgressReporter.cpp">
